core: fix errors after rebasing
