+++ omg/core/views.py
@@ -6,14 +6,11 @@ import arcade
 from omg.entities.events import PickupRequestEvent, ProjectileShotEvent
 from omg.entities.items import Pickupable
 from omg.entities.player import Player
-from omg.entities.projectile import (
-    FireballFactory,
-    IceShardFactory,
-)
 from omg.entities.obstacle import Obstacle
 from omg.mechanics.collision import handle_projectile_collisions
 from omg.mechanics.physics import PhysicsEngineBoundary
 from omg.structural.observer import Observer
+from omg.entities.elements import ELEMENTS
 
 
 ASSET_DIR = os.path.join(
@@ -38,12 +35,15 @@ class GameView(arcade.View):
         self.pickupables: arcade.SpriteList = (
             None  # items that can be pickedup from the ground
         )
+        self.skill_slot_1: arcade.Sprite = None  # Skill slot 1
+        self.skill_slot_2: arcade.Sprite = None  # SKill slot 2
         self.projectiles = None
         self.physics_engine = None
         self.mouse_x = 0
         self.mouse_y = 0
         self.icon_scale = 0.1
-        self.icon_margin = 10
+        self.icon_margin_x = 10
+        self.icon_margin_y = 75
         self.icon_size = 64
         self.active_keys: Dict[tuple, bool] = None
 
@@ -74,6 +74,22 @@ class GameView(arcade.View):
         obstacle.center_y = 300
         self.obstacles.append(obstacle)
 
+        # Create crafted skill slots
+        # Skill slot 1
+        scale_factor_1 = 0.4
+        skill_slot_1_img = os.path.join(ASSET_DIR, "skill_slots_d_f", "D.png")
+        self.skill_slot_1 = arcade.Sprite(skill_slot_1_img, scale=scale_factor_1)
+        self.skill_slot_1.center_x = self.skill_slot_1.width // 2
+        self.skill_slot_1.center_y = self.skill_slot_1.height // 2
+
+        # Skill slot 2
+        scale_factor_2 = 0.4
+        skill_slot_2_img = os.path.join(ASSET_DIR, "skill_slots_d_f", "F.png")
+        self.skill_slot_2 = arcade.Sprite(skill_slot_2_img, scale=scale_factor_2)
+        self.skill_slot_2.center_x = self.skill_slot_1.center_x + \
+            self.skill_slot_2.width
+        self.skill_slot_2.center_y = self.skill_slot_2.height // 2
+
         self.physics_engine = PhysicsEngineBoundary(
             player_sprite=self.player,
             walls=self.obstacles,
@@ -82,32 +98,32 @@ class GameView(arcade.View):
         )
         # Add projectile types
         self.pickupables.append(
-            Pickupable(COIN_IMAGE_PATH, 0.5, FireballFactory, 150, 10)
+            Pickupable(COIN_IMAGE_PATH, 0.5, ELEMENTS["FIRE"], 150, 10)
         )
         self.pickupables.append(
-            Pickupable(COIN_IMAGE_PATH, 0.5, IceShardFactory, 250, 20)
+            Pickupable(COIN_IMAGE_PATH, 0.5, ELEMENTS["ICE"], 250, 20)
         )
         self.pickupables.append(
-            Pickupable(COIN_IMAGE_PATH, 0.5, FireballFactory, 250, 120)
+            Pickupable(COIN_IMAGE_PATH, 0.5, ELEMENTS["FIRE"], 250, 120)
         )
 
     @property
-    def skill_icons(self):
-        """Define self.skill_icons."""
-        # TODO: refactor skill_icons logic if causes a performance bottleneck
-        # For example "skill_acquired event" can trigger an update when necessary
+    def element_icons(self):
+        """Define self.element_icons."""
+        # TODO: refactor element_icons logic if causes a performance bottleneck
+        # For example "element_acquired event" can trigger an update when necessary
         if self.player:
             return [
-                arcade.load_texture(projectile_type.image_file)
-                for projectile_type in self.player.skills
+                arcade.load_texture(element_type["image_file"])
+                for element_type in self.player.elements
             ]
         else:
             return None
 
-    def _load_skill_icons(self):
-        for projectile_type in self.player.skills:
-            icon_texture = arcade.load_texture(projectile_type.image_file)
-            self.skill_icons.append(icon_texture)
+    def _load_element_icons(self):
+        for element_type in self.player.elements:
+            icon_texture = arcade.load_texture(element_type["image_file"])
+            self.element_icons.append(icon_texture)
 
     def on_draw(self):
         """Drawing code."""
@@ -188,10 +204,11 @@ class GameView(arcade.View):
         self.active_keys = {}  # reset the dictionary
 
     def _draw_ui(self):
-        for i, icon in enumerate(self.skill_icons):
-            x = self.icon_margin + i * (self.icon_size + self.icon_margin)
-            y = self.icon_margin
-            if i == self.player.skills.get_current_index():
+        # Draw picked up elements to top left corner
+        for i, icon in enumerate(self.element_icons):
+            x = self.icon_margin_x + i * (self.icon_size + self.icon_margin_x)
+            y = SCREEN_HEIGHT - self.icon_margin_y
+            if i == self.player.elements.get_current_index():
                 arcade.draw_rectangle_outline(
                     x + self.icon_size // 2,
                     y + self.icon_size // 2,
@@ -207,6 +224,35 @@ class GameView(arcade.View):
                 self.icon_size,
                 icon,
             )
+
+        # Draw combined skill icons
+        if self.player.crafted_skill_slots[0] is not None:
+            scale_factor_crafted_skill_1 = 0.3
+            skill_1_img = os.path.join(
+                ASSET_DIR,
+                "skills",
+                f"{self.player.crafted_skill_slots[0]}.png"
+            )
+            skill_1 = arcade.Sprite(skill_1_img, scale=scale_factor_crafted_skill_1)
+            skill_1.center_x = skill_1.width // 2
+            skill_1.center_y = skill_1.height // 2
+            skill_1.draw()
+
+        if self.player.crafted_skill_slots[1] is not None:
+            scale_factor_crafted_skill_2 = 0.3
+            skill_2_img = os.path.join(
+                ASSET_DIR,
+                "skills",
+                f"{self.player.crafted_skill_slots[1]}.png"
+            )
+            skill_2 = arcade.Sprite(skill_2_img, scale=scale_factor_crafted_skill_2)
+            skill_2.center_x = skill_2.width // 2 + skill_2.width
+            skill_2.center_y = skill_2.height // 2
+            skill_2.draw()
+
+        # Draw usable skill slots to bottom left corner
+        self.skill_slot_1.draw()
+        self.skill_slot_2.draw()
 
 
 class PauseView(arcade.View):
